realmate_challenge_back/pyproject.toml
[project]
name = "realmate-challenge"
version = "0.1.0"
description = ""
authors = [{ name = "Seu nome aqui", email = "seu.email@email.com.br" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["django (>=5.1.6,<6.0.0)", "djangorestframework (>=3.15.2,<4.0.0)", "django-cors-headers (>=4.7.0,<5.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
